+++ parallel/include/field.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-bool is_field_literal(int64_t field);
-bool is_field_empty(int64_t field);
+bool is_field_literal(int32_t field);
+bool is_field_empty(int32_t field);
 
-int64_t to_pretty(int64_t field);
-int64_t translate_field(int64_t field);
+int32_t to_pretty(int32_t field);
+int32_t translate_field(int32_t field);
 
 #endif
